+++ src/main/java/org/opensearch/index/store/directio/CryptoDirectIOMemoryIndexInput.java
@@ -4,9 +4,6 @@
  */
 package org.opensearch.index.store.directio;
 
-import static org.opensearch.index.store.directio.DirectIoConfigs.CACHE_BLOCK_MASK;
-import static org.opensearch.index.store.directio.DirectIoConfigs.CACHE_BLOCK_SIZE;
-
 import java.io.EOFException;
 import java.io.IOException;
 import java.lang.foreign.Arena;
@@ -18,7 +15,6 @@ import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -26,14 +22,18 @@ import org.apache.lucene.store.AlreadyClosedException;
 import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.store.RandomAccessInput;
 import org.apache.lucene.util.ArrayUtil;
+import org.opensearch.common.SuppressForbidden;
 import org.opensearch.index.store.block_cache.BlockCache;
-import org.opensearch.index.store.block_cache.BlockCacheKey;
 import org.opensearch.index.store.block_cache.BlockCacheValue;
 import org.opensearch.index.store.block_cache.BlockLoader;
 import org.opensearch.index.store.block_cache.RefCountedMemorySegment;
+import static org.opensearch.index.store.directio.DirectIoConfigs.CACHE_BLOCK_MASK;
+import static org.opensearch.index.store.directio.DirectIoConfigs.CACHE_BLOCK_SIZE;
 
+@SuppressForbidden(reason = "temporary bypass")
 @SuppressWarnings("preview")
 public class CryptoDirectIOMemoryIndexInput extends IndexInput implements RandomAccessInput {
+
     private static final Logger LOGGER = LogManager.getLogger(CryptoDirectIOMemoryIndexInput.class);
 
     static final ValueLayout.OfByte LAYOUT_BYTE = ValueLayout.JAVA_BYTE;
@@ -77,22 +77,44 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
         byte[] key,
         byte[] iv
     ) {
-        return new MultiSegmentImpl(
-            resourceDescription,
-            channel,
-            path,
-            arena,
-            blockCache,
-            blockLoader,
-            segments,
-            refSegments,
-            0L,
-            0L,
-            length,
-            chunkSizePower,
-            key,
-            iv
-        );
+
+        assert Arrays.stream(segments).map(MemorySegment::scope).allMatch(arena.scope()::equals);
+
+        if (segments.length == 1) {
+            return new SingleSegmentImpl(
+                resourceDescription,
+                channel,
+                path,
+                arena,
+                blockCache,
+                blockLoader,
+                segments[0],
+                refSegments,
+                0L,
+                length,
+                chunkSizePower,
+                key,
+                iv
+            );
+        } else {
+            return new MultiSegmentImpl(
+                resourceDescription,
+                channel,
+                path,
+                arena,
+                blockCache,
+                blockLoader,
+                segments,
+                refSegments,
+                0L,
+                0L,
+                length,
+                chunkSizePower,
+                key,
+                iv
+            );
+        }
+
     }
 
     private CryptoDirectIOMemoryIndexInput(
@@ -134,6 +156,16 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
         }
     }
 
+    protected long getAbsoluteBaseOffset() {
+        // Use the existing getFilePointer() which gives us position relative to this input
+        return getFilePointer() + absoluteBaseOffset;
+    }
+
+    protected long getAbsoluteBaseOffset(long pos) {
+        // pos is relative to this input, add base offset for absolute position
+        return pos + absoluteBaseOffset;
+    }
+
     // the unused parameter is just to silence javac about unused variables
     RuntimeException handlePositionalIOOBE(RuntimeException unused, String action, long pos) throws IOException {
         if (pos < 0L) {
@@ -148,143 +180,100 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
         return new AlreadyClosedException("Already closed: " + this);
     }
 
-    private void loadBlockAsync(long fileOffset, int lengthNeeded) {
-        if (lengthNeeded <= 0)
+    private void fillBlockCacheForRange(long fileOffset, long lengthNeeded) throws IOException {
+        if (lengthNeeded == 0)
             return;
 
-        // Calculate EOF for this file/slice
-        long eofOffset = this.absoluteBaseOffset + this.length;
+        long alignedBlockStart = fileOffset & ~CACHE_BLOCK_MASK;
+        long requestEnd = fileOffset + lengthNeeded;
+        long alignedBlockEnd = ((requestEnd + CACHE_BLOCK_SIZE - 1) & ~CACHE_BLOCK_MASK);
 
-        // Calculate which cache blocks we need
-        long startBlockOffset = fileOffset & ~CACHE_BLOCK_MASK;  // Align to cache boundary
-        long endOffset = fileOffset + lengthNeeded;
-        long endBlockOffset = (endOffset + CACHE_BLOCK_MASK) & ~CACHE_BLOCK_MASK;  // Round up
+        int blocksProcessed = 0;
+        long startTime = System.nanoTime();
 
-        // Trigger async loading for all blocks in range (non-blocking)
-        CompletableFuture.runAsync(() -> {
+        for (long blockOffset = alignedBlockStart; blockOffset < alignedBlockEnd; blockOffset += CACHE_BLOCK_SIZE) {
+            DirectIOBlockCacheKey cacheKey = new DirectIOBlockCacheKey(path, blockOffset);
+
             try {
-                // Load all required blocks (only complete blocks)
-                for (long blockOffset = startBlockOffset; blockOffset < endBlockOffset; blockOffset += CACHE_BLOCK_SIZE) {
-                    if (blockOffset + CACHE_BLOCK_SIZE <= eofOffset) { // Only complete blocks
-                        loadCacheBlock(blockOffset);
-                    }
-                }
-            } catch (Exception e) {
-                LOGGER.debug("Background cache loading failed: {}", e.getMessage());
+                blockCache.getOrLoad(cacheKey, CACHE_BLOCK_SIZE, blockLoader);
+                blocksProcessed++;
+            } catch (IOException e) {
+                LOGGER.error("Failed to access cache for block at offset {}: {}", blockOffset, e);
             }
-        });
-
-    }
-
-    private void loadCacheBlock(long blockOffset) {
-        try {
-            BlockCacheKey cacheKey = new DirectIOBlockCacheKey(path, blockOffset);
-            blockCache.getOrLoad(cacheKey, CACHE_BLOCK_SIZE, blockLoader);
-        } catch (IOException e) {
-            LOGGER.debug("Failed to load cache block at offset {}: {}", blockOffset, e.getMessage());
-        }
-    }
-
-    private MemorySegment loadFromBlockCache(int segmentIndex) throws IOException {
-        // Calculate which cache block contains this segment
-        final long segmentOffset = (long) segmentIndex << chunkSizePower;
-        final long absoluteSegmentOffset = this.absoluteBaseOffset + segmentOffset;
-        final long cacheBlockOffset = absoluteSegmentOffset & ~CACHE_BLOCK_MASK;
-
-        final BlockCacheKey cacheKey = new DirectIOBlockCacheKey(path, cacheBlockOffset);
-
-        try {
-            // ONLY get from cache - never load (non-blocking)
-            Optional<BlockCacheValue<RefCountedMemorySegment>> valueOpt = blockCache.get(cacheKey);
-
-            if (valueOpt.isPresent()) {
-                RefCountedMemorySegment refSeg = valueOpt.get().block();
-
-                // Extract our segment from the 128KB cache block
-                long offsetInCacheBlock = absoluteSegmentOffset - cacheBlockOffset;
-                int segmentSize = 1 << chunkSizePower;
-                MemorySegment cacheBlock = refSeg.segment();
-                MemorySegment ourSegment = cacheBlock
-                    .asSlice(offsetInCacheBlock, Math.min(segmentSize, cacheBlock.byteSize() - offsetInCacheBlock));
-
-                segments[segmentIndex] = ourSegment;
-                refSegments[segmentIndex] = refSeg;  // Hold reference
-
-                return ourSegment;
-            }
-        } catch (Exception e) {
-            LOGGER.debug("Failed to access cache for segment at offset {}: {}", absoluteSegmentOffset, e);
         }
 
-        return null;  // Cache miss - that's fine, fall back to mmap
+        long durationMicros = (System.nanoTime() - startTime) / 1_000;
+        LOGGER
+            .info(
+                "Cache fill for {} (offset={}, length={}): blocks={}, took={} us",
+                path.getFileName(),
+                fileOffset,
+                lengthNeeded,
+                blocksProcessed,
+                durationMicros
+            );
     }
 
     // for positioning/seeking (no data needed)
-    private MemorySegment loadSegment(int segmentIndex) throws IOException {
-        return loadSegment(segmentIndex, 0); // 0 = no specific length needed
+    protected MemorySegment loadSegment(int segmentIndex) throws IOException {
+        ensureOpen();
+        // Direct mmap for positioning - no cache interaction
+        return segments[segmentIndex];
     }
 
-    private MemorySegment loadSegment(int segmentIndex, int lengthNeeded) throws IOException {
-
+    protected MemorySegment loadSegment(int segmentIndex, long fileOffset, int lengthNeeded) throws IOException {
         ensureOpen();
 
-        // Trigger async cache population.
-        if (lengthNeeded > 0) {
-            long fileOffset = this.absoluteBaseOffset + ((long) segmentIndex << chunkSizePower) + curPosition;
-
-            if (lengthNeeded > 1024) {
-                LOGGER
-                    .debug(
-                        "Loading segment {} at fileOffset={} (length={} bytes, ~{:.2f} KB)",
-                        segmentIndex,
-                        fileOffset,
-                        lengthNeeded,
-                        lengthNeeded / 1024.0
-                    );
-            } else {
-                LOGGER
-                    .debug(
-                        "Loading small segment {} at fileOffset={} (length={} bytes, ~{:.2f} KB)",
-                        segmentIndex,
-                        fileOffset,
-                        lengthNeeded,
-                        lengthNeeded
-                    );
+        if (lengthNeeded < CACHE_BLOCK_SIZE) {
+            // Small read - always use mmap
+            return segments[segmentIndex];
+        } else {
+            // Check if this segment was already upgraded to cached version
+            if (refSegments[segmentIndex] != null) {
+                // Already cached - no lookup needed!
+                return segments[segmentIndex];
             }
 
-            loadBlockAsync(fileOffset, lengthNeeded);
-        }
+            // Large read - cache-first strategy
+            if (blockCache != null) {
+                long alignedBlockStart = fileOffset & ~CACHE_BLOCK_MASK;
+                DirectIOBlockCacheKey cacheKey = new DirectIOBlockCacheKey(path, alignedBlockStart);
 
-        // Try cache (non-blocking)
-        // the assumption is all reads will fit in one block.
-        MemorySegment cachedSegment = loadFromBlockCache(segmentIndex);
-        if (cachedSegment != null) {
-            return cachedSegment;
-        }
+                Optional<BlockCacheValue<RefCountedMemorySegment>> cached = blockCache.get(cacheKey);
 
-        // Fall back to reliable mmap (immediate, always works)
-        return segments[segmentIndex];
+                if (cached.isPresent()) {
+                    RefCountedMemorySegment refSeg = cached.get().block();
+                    segments[segmentIndex] = refSeg.segment();
+                    refSegments[segmentIndex] = refSeg;  // Mark as upgraded
+
+                    return segments[segmentIndex];
+                }
+            }
+
+            return segments[segmentIndex];  // fallback to mmap
+        }
     }
 
     @Override
     public final byte readByte() throws IOException {
         try {
-            curSegment = loadSegment(curSegmentIndex, 1);
-            byte v = curSegment.get(LAYOUT_BYTE, curPosition);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, 1);
+            final byte v = curSegment.get(LAYOUT_BYTE, curPosition);
             curPosition++;
             return v;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException e) {
-            // Move to next segment
             do {
                 curSegmentIndex++;
                 if (curSegmentIndex >= segments.length) {
                     throw new EOFException("read past EOF: " + this);
                 }
-                curSegment = loadSegment(curSegmentIndex, 1);
+
+                long fileOffset = getAbsoluteBaseOffset();
+                curSegment = loadSegment(curSegmentIndex, fileOffset, 1);
                 curPosition = 0L;
             } while (curSegment.byteSize() == 0L);
-
-            byte v = curSegment.get(LAYOUT_BYTE, curPosition);
+            final byte v = curSegment.get(LAYOUT_BYTE, curPosition);
             curPosition++;
             return v;
         } catch (NullPointerException | IllegalStateException e) {
@@ -295,20 +284,27 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     @Override
     public final void readBytes(byte[] b, int offset, int len) throws IOException {
         try {
-            curSegment = loadSegment(curSegmentIndex, len);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, len);
             MemorySegment.copy(curSegment, LAYOUT_BYTE, curPosition, b, offset, len);
             curPosition += len;
-        } catch (@SuppressWarnings("unused") IndexOutOfBoundsException e) {
+        } catch (IndexOutOfBoundsException e) {
             readBytesBoundary(b, offset, len);
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
+        } catch (IOException e) {
+            throw new IOException("Decryption or read failed", e);
         }
     }
 
     private void readBytesBoundary(byte[] b, int offset, int len) throws IOException {
+        long startFileOffset = getAbsoluteBaseOffset();
+        int originalLen = len;
         try {
             long curAvail = curSegment.byteSize() - curPosition;
             while (len > curAvail) {
+                long fileOffset = startFileOffset + (originalLen - len);
+                curSegment = loadSegment(curSegmentIndex, fileOffset, len);
                 MemorySegment.copy(curSegment, LAYOUT_BYTE, curPosition, b, offset, (int) curAvail);
                 len -= curAvail;
                 offset += curAvail;
@@ -316,23 +312,27 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
                 if (curSegmentIndex >= segments.length) {
                     throw new EOFException("read past EOF: " + this);
                 }
-                curSegment = segments[curSegmentIndex];
-                curSegment = loadSegment(curSegmentIndex, len);
                 curPosition = 0L;
                 curAvail = curSegment.byteSize();
             }
+
+            long fileOffset = startFileOffset + (originalLen - len);
+            curSegment = loadSegment(curSegmentIndex, fileOffset, len);
             MemorySegment.copy(curSegment, LAYOUT_BYTE, curPosition, b, offset, len);
             curPosition += len;
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
+        } catch (IOException e) {
+            throw new IOException("Decryption failed", e);
         }
     }
 
     private void ensureSegmentsForBoundaryCrossing(int elementCount, int elementSize) throws IOException {
         long totalBytes = elementSize * (long) elementCount;
-        curSegment = loadSegment(curSegmentIndex, (int) totalBytes);
-        long currentRemaining = curSegment.byteSize() - curPosition;
+        long fileOffset = absoluteBaseOffset + ((long) curSegmentIndex << chunkSizePower) + curPosition;
+        curSegment = loadSegment(curSegmentIndex, fileOffset, (int) totalBytes);
 
+        long currentRemaining = curSegment.byteSize() - curPosition;
         // If read spans beyond current segment, ensure next segments are loaded
         if (totalBytes > currentRemaining) {
             long remainingBytes = totalBytes - currentRemaining;
@@ -340,7 +340,8 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
 
             for (int i = 1; i <= segmentsToLoad && (curSegmentIndex + i) < segments.length; i++) {
                 long bytesForThisSegment = Math.min(remainingBytes, 1L << chunkSizePower);
-                loadSegment(curSegmentIndex + i, (int) bytesForThisSegment);
+                long nextSegmentFileOffset = absoluteBaseOffset + ((long) (curSegmentIndex + i) << chunkSizePower);
+                loadSegment(curSegmentIndex + i, nextSegmentFileOffset, (int) bytesForThisSegment);
                 remainingBytes -= bytesForThisSegment;
             }
         }
@@ -350,10 +351,13 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     public void readInts(int[] dst, int offset, int length) throws IOException {
         try {
             int totalBytes = Integer.BYTES * length;
-            curSegment = loadSegment(curSegmentIndex, totalBytes);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, totalBytes);
             MemorySegment.copy(curSegment, LAYOUT_LE_INT, curPosition, dst, offset, length);
-            curPosition += Integer.BYTES * (long) length;
+            curPosition += totalBytes;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException iobe) {
+            // Crossing segment boundaries - decrypt current segment remainder and next segment
+            // Decrypt remainder of current segment
             ensureSegmentsForBoundaryCrossing(length, Integer.BYTES);
             super.readInts(dst, offset, length);
         } catch (NullPointerException | IllegalStateException e) {
@@ -364,11 +368,15 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     @Override
     public void readLongs(long[] dst, int offset, int length) throws IOException {
         try {
+            // Decrypt the range we're about to read
             int totalBytes = Long.BYTES * length;
-            curSegment = loadSegment(curSegmentIndex, totalBytes);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, totalBytes);
+
             MemorySegment.copy(curSegment, LAYOUT_LE_LONG, curPosition, dst, offset, length);
-            curPosition += Long.BYTES * (long) length;
+            curPosition += totalBytes;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException iobe) {
+            // Crossing segment boundaries - decrypt segments then delegate to super
             ensureSegmentsForBoundaryCrossing(length, Long.BYTES);
             super.readLongs(dst, offset, length);
         } catch (NullPointerException | IllegalStateException e) {
@@ -380,10 +388,13 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     public void readFloats(float[] dst, int offset, int length) throws IOException {
         try {
             int totalBytes = Float.BYTES * length;
-            curSegment = loadSegment(curSegmentIndex, totalBytes);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, totalBytes);
+
             MemorySegment.copy(curSegment, LAYOUT_LE_FLOAT, curPosition, dst, offset, length);
-            curPosition += Float.BYTES * (long) length;
+            curPosition += totalBytes;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException iobe) {
+            // Crossing segment boundaries - decrypt segments then delegate to super
             ensureSegmentsForBoundaryCrossing(length, Float.BYTES);
             super.readFloats(dst, offset, length);
         } catch (NullPointerException | IllegalStateException e) {
@@ -394,16 +405,14 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     @Override
     public final short readShort() throws IOException {
         try {
-            curSegment = loadSegment(curSegmentIndex, Short.BYTES);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, Short.BYTES);
             final short v = curSegment.get(LAYOUT_LE_SHORT, curPosition);
             curPosition += Short.BYTES;
             return v;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException e) {
-            // Any single primitive read (short/int/long) can span at most 2 segments
-            // when crossing a boundary, so we only need to ensure next segment is loaded
-            if (curSegmentIndex + 1 < segments.length) {
-                loadSegment(curSegmentIndex + 1, Short.BYTES);
-            }
+            ensureSegmentsForBoundaryCrossing(1, Short.BYTES);
+
             return super.readShort();
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
@@ -413,14 +422,14 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     @Override
     public final int readInt() throws IOException {
         try {
-            curSegment = loadSegment(curSegmentIndex, Integer.BYTES);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, Integer.BYTES);
             final int v = curSegment.get(LAYOUT_LE_INT, curPosition);
             curPosition += Integer.BYTES;
             return v;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException e) {
-            if (curSegmentIndex + 1 < segments.length) {
-                loadSegment(curSegmentIndex + 1, Integer.BYTES);
-            }
+            // Crossing segment boundaries - ensure segments are loaded then delegate to super
+            ensureSegmentsForBoundaryCrossing(1, Integer.BYTES);
             return super.readInt();
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
@@ -430,14 +439,14 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     @Override
     public final long readLong() throws IOException {
         try {
-            curSegment = loadSegment(curSegmentIndex, Long.BYTES);
+            long fileOffset = getAbsoluteBaseOffset();
+            curSegment = loadSegment(curSegmentIndex, fileOffset, Long.BYTES);
             final long v = curSegment.get(LAYOUT_LE_LONG, curPosition);
             curPosition += Long.BYTES;
             return v;
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException e) {
-            if (curSegmentIndex + 1 < segments.length) {
-                loadSegment(curSegmentIndex + 1, Long.BYTES);
-            }
+            ensureSegmentsForBoundaryCrossing(1, Short.BYTES);
+
             return super.readLong();
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
@@ -487,8 +496,10 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     public byte readByte(long pos) throws IOException {
         try {
             final int si = (int) (pos >> chunkSizePower);
-            MemorySegment segment = loadSegment(si, 1);
-            return segment.get(LAYOUT_BYTE, pos & chunkSizeMask);
+            final long segmentOffset = pos & chunkSizeMask;
+            long fileOffset = getAbsoluteBaseOffset(pos);
+            MemorySegment segment = loadSegment(si, fileOffset, 1);
+            return segment.get(LAYOUT_BYTE, segmentOffset);
         } catch (IndexOutOfBoundsException ioobe) {
             throw handlePositionalIOOBE(ioobe, "read", pos);
         } catch (NullPointerException | IllegalStateException e) {
@@ -499,45 +510,60 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     @Override
     public short readShort(long pos) throws IOException {
         final int si = (int) (pos >> chunkSizePower);
+        final long segmentOffset = pos & chunkSizeMask;
         try {
-            MemorySegment segment = loadSegment(si, Short.BYTES);
-            return segment.get(LAYOUT_LE_SHORT, pos & chunkSizeMask);
+            long fileOffset = getAbsoluteBaseOffset(pos);
+            MemorySegment segment = loadSegment(si, fileOffset, Short.BYTES);
+            return segment.get(LAYOUT_LE_SHORT, segmentOffset);
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException ioobe) {
             // either it's a boundary, or read past EOF, fall back:
-            setPos(pos, si);  // This will load the segment
-            return readShort(); // Sequential readShort() already handles lazy loading
+            setPos(pos, si);
+            return readShort();
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
         }
+
     }
 
     @Override
     public int readInt(long pos) throws IOException {
         final int si = (int) (pos >> chunkSizePower);
+        final long segmentOffset = pos & chunkSizeMask;
+
         try {
-            MemorySegment segment = loadSegment(si, Integer.BYTES);
-            return segment.get(LAYOUT_LE_INT, pos & chunkSizeMask);
+            // Add decryption before reading
+            long fileOffset = getAbsoluteBaseOffset(pos);
+            MemorySegment segment = loadSegment(si, fileOffset, Integer.BYTES);
+            return segment.get(LAYOUT_LE_INT, segmentOffset);
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException ioobe) {
             // either it's a boundary, or read past EOF, fall back:
             setPos(pos, si);
             return readInt();
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
+        } catch (IOException e) {
+            throw new IOException("Decryption failed", e);
         }
     }
 
     @Override
     public long readLong(long pos) throws IOException {
         final int si = (int) (pos >> chunkSizePower);
+        final long segmentOffset = pos & chunkSizeMask;
+
         try {
-            MemorySegment segment = loadSegment(si, Long.BYTES);
-            return segment.get(LAYOUT_LE_LONG, pos & chunkSizeMask);
+            // Add decryption before reading
+            long fileOffset = getAbsoluteBaseOffset(pos);
+            MemorySegment segment = loadSegment(si, fileOffset, Long.BYTES);
+            return segment.get(LAYOUT_LE_LONG, segmentOffset);
         } catch (@SuppressWarnings("unused") IndexOutOfBoundsException ioobe) {
             // either it's a boundary, or read past EOF, fall back:
             setPos(pos, si);
             return readLong();
         } catch (NullPointerException | IllegalStateException e) {
             throw alreadyClosed(e);
+        } catch (IOException e) {
+            throw new IOException("Decryption failed", e);
         }
     }
 
@@ -559,8 +585,8 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
     }
 
     /**
-     * Creates a slice of this index input, with the given description, offset, and length. The slice
-     * is seeked to the beginning.
+     * Creates a slice of this index input, with the given description, offset,
+     * and length. The slice is seeked to the beginning.
      */
     @Override
     public final CryptoDirectIOMemoryIndexInput slice(String sliceDescription, long offset, long length) {
@@ -582,41 +608,88 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
         return buildSlice(sliceDescription, offset, length);
     }
 
+    /**
+     * Builds the actual sliced IndexInput (may apply extra offset in
+     * subclasses). *
+     */
     CryptoDirectIOMemoryIndexInput buildSlice(String sliceDescription, long offset, long length) {
         ensureOpen();
+
+        // Calculate the absolute file position where this slice starts
+        // This is crucial for decryption - we need to know where in the original file we are
+        final long sliceAbsoluteOffset = this.absoluteBaseOffset + offset;
 
         final long sliceEnd = offset + length;
         final int startIndex = (int) (offset >>> chunkSizePower);
         final int endIndex = (int) (sliceEnd >>> chunkSizePower);
 
-        // we always allocate one more slice, the last one may be a 0 byte one after truncating with
-        // asSlice():
-        final MemorySegment slices[] = ArrayUtil.copyOfSubArray(segments, startIndex, endIndex + 1);
+        // Ensure we don't go beyond array bounds
+        final int actualEndIndex = Math.min(endIndex, segments.length - 1);
+        // Copy the segments to prepare our view.
+        final MemorySegment slices[] = ArrayUtil.copyOfSubArray(segments, startIndex, actualEndIndex + 1);
 
-        // set the last segment's limit for the sliced view.
-        slices[slices.length - 1] = slices[slices.length - 1].asSlice(0L, sliceEnd & chunkSizeMask);
+        // Set the last segment's limit for the sliced view
 
-        final long sliceOffset = offset & chunkSizeMask;
+        if (slices.length > 0) {
+            long lastSegmentLimit = sliceEnd & chunkSizeMask;
+            if (lastSegmentLimit > 0) {
+                slices[slices.length - 1] = slices[slices.length - 1].asSlice(0L, lastSegmentLimit);
+            }
+        }
+
+        // Convert offset to position within the first segment
+        final long segmentOffset = offset & chunkSizeMask;
+
+        LOGGER
+            .debug(
+                "Building slice: description={}, parentOffset={}, length={}, " + "absoluteFileOffset={}, segmentOffset={}, startIndex={}",
+                sliceDescription,
+                offset,
+                length,
+                sliceAbsoluteOffset,
+                segmentOffset,
+                startIndex
+            );
+
         final String newResourceDescription = getFullSliceDescription(sliceDescription);
 
-        final long sliceAbsoluteOffset = this.absoluteBaseOffset + offset;  // Simple addition!
+        if (slices.length == 1) {
 
-        return new MultiSegmentImpl(
-            newResourceDescription,
-            null,  // channel
-            path,
-            null,  // arena
-            blockCache,
-            blockLoader,
-            slices,
-            null, // reference segment tracking null for slices.
-            sliceOffset,
-            sliceAbsoluteOffset,
-            length,
-            chunkSizePower,
-            key,
-            iv
-        );
+            return new SingleSegmentImpl(
+                newResourceDescription,
+                null,  // channel
+                path,
+                null,  // arena
+                blockCache,
+                blockLoader,
+                slices[0].asSlice(segmentOffset, length),
+                null, // reference segment tracking null for slices.
+                sliceAbsoluteOffset,
+                length,
+                chunkSizePower,
+                key,
+                iv
+            );
+
+        } else {
+            return new MultiSegmentImpl(
+                newResourceDescription,
+                null,  // channel
+                path,
+                null,  // arena
+                blockCache,
+                blockLoader,
+                slices,
+                null, // reference segment tracking null for slices.
+                segmentOffset,
+                sliceAbsoluteOffset,
+                length,
+                chunkSizePower,
+                key,
+                iv
+            );
+        }
+
     }
 
     @Override
@@ -666,7 +739,120 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
         isOpen = false;
     }
 
-    /** This class adds offset support to MemorySegmentIndexInput, which is needed for slices. */
+    /**
+     * Optimization of MemorySegmentIndexInput for when there is only one
+     * segment.
+     */
+    static final class SingleSegmentImpl extends CryptoDirectIOMemoryIndexInput {
+        SingleSegmentImpl(
+            String resourceDescription,
+            FileChannel channel,
+            Path path,
+            Arena arena,
+            BlockCache<RefCountedMemorySegment> blockCache,
+            BlockLoader<RefCountedMemorySegment> blockLoader,
+            MemorySegment segment,
+            RefCountedMemorySegment[] refSegments,
+            long absoluteBaseOffset,    // absolute file offset of first segment
+            long length,
+            int chunkSizePower,
+            byte[] key,
+            byte[] iv
+        ) {
+            super(
+                resourceDescription,
+                channel,
+                path,
+                arena,
+                blockCache,
+                blockLoader,
+                new MemorySegment[] { segment },
+                refSegments,
+                absoluteBaseOffset,
+                length,
+                chunkSizePower,
+                key,
+                iv
+            );
+            this.curSegmentIndex = 0;
+        }
+
+        @Override
+        public void seek(long pos) throws IOException {
+            ensureOpen();
+            try {
+                curPosition = Objects.checkIndex(pos, length + 1);
+            } catch (IndexOutOfBoundsException e) {
+                throw handlePositionalIOOBE(e, "seek", pos);
+            }
+        }
+
+        @Override
+        public long getFilePointer() {
+            ensureOpen();
+            return curPosition;
+        }
+
+        @Override
+        public byte readByte(long pos) throws IOException {
+            try {
+                // For single segment, pos is the absolute file position
+                curSegment = super.loadSegment(0, getAbsoluteBaseOffset(pos), 1);
+                return curSegment.get(LAYOUT_BYTE, pos);
+            } catch (IndexOutOfBoundsException e) {
+                throw handlePositionalIOOBE(e, "read", pos);
+            } catch (NullPointerException | IllegalStateException e) {
+                throw alreadyClosed(e);
+            }
+        }
+
+        @Override
+        public short readShort(long pos) throws IOException {
+            try {
+                curSegment = super.loadSegment(0, getAbsoluteBaseOffset(pos), 2);
+                return curSegment.get(LAYOUT_LE_SHORT, pos);
+            } catch (IndexOutOfBoundsException e) {
+                throw handlePositionalIOOBE(e, "read", pos);
+            } catch (NullPointerException | IllegalStateException e) {
+                throw alreadyClosed(e);
+            }
+        }
+
+        @Override
+        public int readInt(long pos) throws IOException {
+            try {
+                // Decrypt 4 bytes for int
+                long addr = curSegment.address() + pos;
+                curSegment = super.loadSegment(0, getAbsoluteBaseOffset(pos), 4);
+                return curSegment.get(LAYOUT_LE_INT, pos);
+            } catch (IndexOutOfBoundsException e) {
+                throw handlePositionalIOOBE(e, "read", pos);
+            } catch (NullPointerException | IllegalStateException e) {
+                throw alreadyClosed(e);
+            } catch (IOException e) {
+                throw new IOException("Decryption failed", e);
+            }
+        }
+
+        @Override
+        public long readLong(long pos) throws IOException {
+            try {
+                curSegment = super.loadSegment(0, getAbsoluteBaseOffset(pos), 8);
+                return curSegment.get(LAYOUT_LE_LONG, pos);
+            } catch (IndexOutOfBoundsException e) {
+                throw handlePositionalIOOBE(e, "read", pos);
+            } catch (NullPointerException | IllegalStateException e) {
+                throw alreadyClosed(e);
+            } catch (IOException e) {
+                throw new IOException("Decryption failed", e);
+            }
+        }
+    }
+
+    /**
+     * This class adds offset support to MemorySegmentIndexInput, which is
+     * needed for slices.
+     */
     static final class MultiSegmentImpl extends CryptoDirectIOMemoryIndexInput {
         private final long offset;
 
@@ -680,7 +866,7 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
             MemorySegment[] segments,
             RefCountedMemorySegment[] refSegments,
             long offset,
-            long absoluteBaseOffset,    // absolute file offset of first segment
+            long absoluteBaseOffset,
             long length,
             int chunkSizePower,
             byte[] key,
@@ -702,22 +888,12 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
                 iv
             );
             this.offset = offset;
-
-            LOGGER
-                .info(
-                    "MultiSegmentImpl created: offset={}, absoluteBaseOffset={}, length={}, parent.length={}",
-                    offset,
-                    absoluteBaseOffset,
-                    length,
-                    super.length()
-                );
-
             try {
                 seek(0L);
             } catch (IOException ioe) {
                 throw new AssertionError(ioe);
             }
-            assert curSegmentIndex >= 0;
+            assert curSegment != null && curSegmentIndex >= 0;
         }
 
         @Override
@@ -759,6 +935,18 @@ public class CryptoDirectIOMemoryIndexInput extends IndexInput implements Random
         @Override
         CryptoDirectIOMemoryIndexInput buildSlice(String sliceDescription, long ofs, long length) {
             return super.buildSlice(sliceDescription, this.offset + ofs, length);
+        }
+
+        @Override
+        protected long getAbsoluteBaseOffset() {
+            // getFilePointer() already returns position relative to this slice
+            return absoluteBaseOffset + offset + super.getFilePointer();
+        }
+
+        @Override
+        protected long getAbsoluteBaseOffset(long pos) {
+            // pos is relative to the slice, we need to add offset
+            return absoluteBaseOffset + offset + pos;
         }
     }
 }
